Add files via upload
Salib was fired from the programming job due to excessive boredom. After a lot of effort, the house of Kari is struggling with a drug for himself.

He was assigned to accounting tasks related to pharmacies. But since he knows programming; It has been decided that he will write a program to do this, but because he is too bored, he asks you to write it for him.